move validate to component folder, clean up naming
