+++ lib/components/stand-form/__tests__/validate.test.ts
@@ -1,0 +1,3 @@
+/**
+ * Tests to verify stand form validation functions as expected.
+ */
